contributing/samples/monitor_processual/pyproject.toml
[build-system]
requires = ["flit_core>=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "monitor-processual"
version = "0.2.0"
description = "Agente de monitoramento de movimentacoes processuais usando Google ADK"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
monitor-processual = "monitor_processual.main:cli"

[tool.flit.module]
name = "monitor_processual"
